Update credit balance

+++ server/src/db/migrations/create/createCreditsTable.sql
@@ -1,6 +1,6 @@
 CREATE TABLE credits (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
-    balance INT DEFAULT 10,
+    balance NUMERIC(10, 2) DEFAULT 10,
     last_updated TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     user_id UUID NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
